requirements.txt
streamlit
groq
kokoro-onnx
soundfile
requests
beautifulsoup4
PyMuPDF
google-generativeai
